# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateQcOnlineRequest:
    properties:
      details:
        items:
          $ref: '#/definitions/controllers.QcOnlineDetailRequest'
        type: array
      tracking:
        example: TRK123456
        type: string
    required:
    - details
    - tracking
    type: object
...
  controllers.QcOnlineDetailRequest:
    properties:
      box_id:
        example: 1
        type: integer
      quantity:
        example: 5
        minimum: 1
        type: integer
    required:
    - box_id
    - quantity
    type: object
...
  controllers.QcOnlinesListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utils.PaginationResponse'
      qc_onlines:
        items:
          $ref: '#/definitions/models.QcOnlineResponse'
        type: array
    type: object
...
  models.BoxResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.OrderDetailResponse:
    properties:
      id:
        type: integer
      product:
        $ref: '#/definitions/models.ProductResponse'
      product_name:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      variant:
        type: string
    type: object
  models.OrderResponse:
    properties:
      buyer:
        type: string
      channel:
        type: string
      complained:
        type: boolean
      courier:
        type: string
      created_at:
        type: string
      id:
        type: integer
      order_details:
        items:
          $ref: '#/definitions/models.OrderDetailResponse'
        type: array
      order_id:
        type: string
      picked_at:
        type: string
      picked_by:
        type: string
      status:
        type: string
      store:
        type: string
      tracking:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ProductResponse:
    properties:
      barcode:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      name:
        type: string
      sku:
        type: string
      updated_at:
        type: string
      variant:
        type: string
    type: object
  models.QcOnlineDetailResponse:
    properties:
      box:
        $ref: '#/definitions/models.BoxResponse'
      box_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      qc_online_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.QcOnlineResponse:
    properties:
      complained:
        type: boolean
      created_at:
        type: string
      details:
        items:
          $ref: '#/definitions/models.QcOnlineDetailResponse'
        type: array
      id:
        type: integer
      order:
        allOf:
        - $ref: '#/definitions/models.OrderResponse'
        description: Related data
      tracking:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
      user_id:
        type: integer
    type: object
...
  models.RoleResponse:
    properties:
      assigned_at:
        type: string
      assigned_by:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      roles:
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: 192.168.31.136:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/qc-onlines:
    get:
      consumes:
      - application/json
      description: Get list of qc-onlines for current user filtered by current date
        (logged-in users only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by tracking number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.QcOnlinesListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all qc-onlines for logged-in user
      tags:
      - qc-onlines
    post:
      consumes:
      - application/json
      description: Create new qc-online entry with multiple box details (logged-in
        users only)
      parameters:
      - description: Create qc-online request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateQcOnlineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.QcOnlineResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new qc-online
      tags:
      - qc-onlines
  /api/qc-onlines/{id}:
    get:
      consumes:
      - application/json
      description: Get  a specific qc-online by ID (logged-in users only)
      parameters:
      - description: QcOnline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.QcOnlineResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get a specific qc-online by ID
      tags:
      - qc-onlines
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abcdef12345'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"